requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-on-heroku==1.1.2
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.2.0
